requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
billiard==4.2.1
black==24.8.0
build==1.3.0
celery==5.4.0
cffi==1.17.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.12.0
cron-descriptor==1.4.5
cryptography==44.0.1
dj-database-url==2.2.0
Django==5.1.1
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cleanup==9.0.0
django-cors-headers==4.4.0
django-filter==24.3
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
flower==2.0.1
gunicorn==23.0.0
humanize==4.12.1
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
pip-tools==7.5.1
platformdirs==4.3.6
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
pyproject_hooks==1.2.0
pytest==8.3.3
pytest-cov==7.0.0
pytest-django==4.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.1
PyYAML==6.0.2
redis==5.1.1
referencing==0.35.1
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.1
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
